requirements.txt
pystray>=0.19
Pillow>=10.0
customtkinter>=5.2
numpy>=1.24
sounddevice>=0.4
yt-dlp>=2024.1
